# cat-service/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Welcome to cat-service.",
      "version": "1.0.0",
      "title": "cat-service"
    },
...
    "schemes": [
        "http"
    ],
    "paths": {
        "/devices": {
            "get": {
                "tags": [
                    "devices"
                ],
                "description": "Fetch devices",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "handle",
                        "in": "query",
                        "description": "Filter by handle",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "afterDate",
                        "in": "query",
                        "description": "Fetch devices created after the provided date",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "beforeDate",
                        "in": "query",
                        "description": "Fetch devices created before the provided date",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found devices",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/device"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "devices"
                ],
                "description": "Create device",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "handle",
                        "in": "formData",
                        "description": "Used to identify device",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "Device description",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "location",
                        "in": "formData",
                        "description": "Used to help locate device",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ipAddress",
                        "in": "formData",
                        "description": "Device's IP address",
                        "required": false,
                        "type": "string",
                        "format": "ipv4"
                    },
                    {
                        "name": "settings",
                        "in": "formData",
                        "description": "Device settings",
                        "required": false,
                        "type": "string",
                        "format": "json"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Device created",
                        "schema": {
                            "$ref": "#/definitions/device"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/devices/{id}": {
            "get": {
                "tags": [
                    "devices"
                ],
                "description": "Fetch device",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found device",
                        "schema": {
                            "$ref": "#/definitions/device"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "devices"
                ],
                "description": "Update device",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "handle",
                        "in": "formData",
                        "description": "Used to identify device",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "Device description",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "location",
                        "in": "formData",
                        "description": "Used to help locate device",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ipAddress",
                        "in": "formData",
                        "description": "Device's IP address",
                        "required": false,
                        "type": "string",
                        "format": "ipv4"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device updated",
                        "schema": {
                            "$ref": "#/definitions/device"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "devices"
                ],
                "description": "Delete device",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Device deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/devices/{id}/settings": {
            "get": {
                "tags": [
                    "settings"
                ],
                "description": "Fetch device settings by device id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found device settings",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/device-settings"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "settings"
                ],
                "description": "Update device settings by id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "settings",
                        "in": "formData",
                        "description": "Device settings",
                        "required": true,
                        "type": "string",
                        "format": "json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated device settings",
                        "schema": {
                            "$ref": "#/definitions/device-settings"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/devices/{id}/recordings": {
            "get": {
                "tags": [
                    "recordings"
                ],
                "description": "Fetch recordings by device id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found recordings",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/recording"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/devices/{id}/speakers": {
            "get": {
                "tags": [
                    "speakers"
                ],
                "description": "Fetch speakers by device id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found speakers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/speaker"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "device": {
            "type": "object",
            "required": [
                "deviceId",
                "handle",
                "createdDate"
            ],
            "properties": {
                "deviceId": {
                    "type": "string",
                    "format": "uuid"
                },
                "handle": {
                    "type": "string"
                },
                "createdDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string",
                    "format": "ipv4"
                },
                "settings": {
                    "$ref": "#/definitions/device-settings"
                }
            }
        },
        "device-settings": {
            "type": "object",
            "required": [
                "settingsId",
                "deviceId",
                "createdDate"
            ],
            "properties": {
                "settingsId": {
                    "type": "string",
                    "format": "uuid"
                },
                "deviceId": {
                    "type": "string",
                    "format": "uuid"
                },
                "createdDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "properties": {
                    "type": "object"
                }
            }
        },
        "recording": {
            "type": "object",
            "required": [
                "recordingId",
                "recordingTime"
            ],
            "properties": {
                "recordingId": {
                    "type": "string",
                    "format": "uuid"
                },
                "deviceId": {
                    "type": "string",
                    "format": "uuid"
                },
                "speakerId": {
                    "type": "string",
                    "format": "uuid"
                },
                "settingsId": {
                    "type": "string",
                    "format": "uuid"
                },
                "recordingTime": {
                    "type": "string",
                    "format": "date-time"
                },
                "data": {
                    "type": "object"
                }
            }
        },
        "speaker": {
            "type": "object",
            "required": [
                "speakerId",
                "createdDate"
            ],
            "properties": {
                "speakerId": {
                    "type": "string",
                    "format": "uuid"
                },
                "deviceId": {
                    "type": "string",
                    "format": "uuid"
                },
                "createdDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "data": {
                    "type": "object"
                }
            }
        }
    }
}